feat: :sparkles: add Commitizen version bump with GitHub workflow (#71)
# Description

Adding auto-version update workflow via Commitizen. Right now though, I
don't know if the `version_files` will work though, since it is basic
"search and replace for version number in the given files". The
`datapackage.json` file will/may have two `version` fields, one in the
top level and another in the `sources` level. So this might also bump
those versions too (like we had an issue with in Sprout). So we'll have
to see how it goes.

Closes #51
Closes #55

This PR needs a quick review.

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>