feat: :sparkles: workflow to check for updates and make a PR with them (#135)
# Description

This adds a workflow to the created template that will run every day to
check for updates from the template. If there is an update, it will make
a pull request of those changes.

I don't know if it will work though, so we'll have to see how it works
in the data packages that use this template.

Closes #38

This PR needs an in-depth review.

## Checklist

- [x] Ran `just run-all`